+++ sampledb/models/migrations/utils.py
@@ -245,7 +245,8 @@ def get_migrations() -> typing.List[typing.Tuple[int, str, typing.Callable[[typi
         "instrument_log_file_attachments_generate_image_infos",
         "fed_identity_add_login_column",
         "components_add_fed_login_available_column",
-        "authentication_type_add_federated_login"
+        "authentication_type_add_federated_login",
+        "locations_add_enable_object_assignments",
     ]
 
     migrations = []